refactor: 패키지 정리 및 불필요한 클래스 제거
